requirements.txt
altgraph @ file:///home/conda/feedstock_root/build_artifacts/altgraph_1632761154050/work
certifi==2021.10.8
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
fonttools @ file:///D:/bld/fonttools_1643722825909/work
future @ file:///D:/bld/future_1635819659513/work
kiwisolver @ file:///D:/bld/kiwisolver_1635836839743/work
macholib @ file:///home/conda/feedstock_root/build_artifacts/macholib_1595268426755/work
matplotlib @ file:///D:/bld/matplotlib-suite_1639359275746/work
munkres==1.1.4
numpy @ file:///D:/bld/numpy_1643959091469/work
olefile==0.46
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pefile @ file:///home/conda/feedstock_root/build_artifacts/pefile_1630688703091/work
Pillow @ file:///C:/ci/pillow_1602770972588/work
pycryptodome @ file:///D:/bld/pycryptodome_1644028956458/work
pyinstaller @ file:///C:/ci/pyinstaller_1645412385847/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1642753572664/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pywin32==303
pywin32-ctypes @ file:///D:/bld/pywin32-ctypes_1635826703191/work
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tornado @ file:///D:/bld/tornado_1635819758615/work
unicodedata2 @ file:///D:/bld/unicodedata2_1640031186081/work
wincertstore==0.2
